# SNOW-2032053: Added logic to propagate trace context, if it exists, to GS