### Description

SNOW-2032053 

Added support for trace propagation by using OpenTelemetry's trace context injector for Go. This change adds one additional dependency for the Go driver, which is OpenTelemetry v1.35.0. 

The only functional change is an extra couple headers are injected into the HTTPS request to GS, and GS is already equipped to extract it on the other side if it is present - i.e. the header "traceparent", where the value is in W3 Trace Context Format.

This change will allow trace context to be propagated from anywhere the Go driver is being used to GS. The primary use case is for a Go service running on SPCS. The user will then be able to see the entire line of execution in Snowsight in one view, from their SPCS Service to the SQL queries they ran using the connector. 

This is a feature request from Relational AI, an external customer.

The compilation time and size as well as the runtime were benchmarked before and after the change using the `cm/select1` program. These were tested (built and ran) 5 times each.

Compilation/Runtime for cmd/select1 before/after these changes:
- Before changes (times are averaged):
  - Compilation Time: `2.21 seconds`
  - Runtime: `0.8 seconds`
  - Compilation Size: `33.35MB`
 
- With changes (times are averaged):
  - Compilation Time: `2.25 seconds`
  - Runtime: `0.82 seconds`
  - Compilation Size: `33.43MB`

### Checklist
- [x] Created tests which fail without the change (if possible)
- [x] Extended the README / documentation, if necessary