@sfc-gh-snowflakedb-snyk-sa 2025-04-09
![](https://res.cloudinary.com/snyk/image/upload/r-d/scm-platform/snyk-pull-requests/pr-banner-default.svg)

### :tada: **Snyk checks have passed. No issues have been found so far.**

:white_check_mark: **security/snyk** check is complete. No issues have been found. [(View Details)](https://app.snyk.io/org/snowflakedb-sca-scanning-public-repo/pr-checks/6c5311f5-eb4d-4f78-9bdb-2d4e9b79a9f0?refs=5a7bcca8d62a006f5529171cb5ddccfc4554f5d5&source=prChecksComment)

:white_check_mark: **license/snyk** check is complete. No issues have been found. [(View Details)](https://app.snyk.io/org/snowflakedb-sca-scanning-public-repo/pr-checks/6c5311f5-eb4d-4f78-9bdb-2d4e9b79a9f0/license?refs=5a7bcca8d62a006f5529171cb5ddccfc4554f5d5&source=prChecksComment)